Moved `MirrordPolicy` to a separate module, added `MirrordClusterPolicy` (#2974)
* Moved MirrordPolicy to a separate module, added MirrordClusterPolicy

* Doc and test fix
