+++ changelog.d/+cluster-policy.added.md
@@ -1,0 +1,1 @@
+Added custom resource definition for cluster-wide mirrord policy - `MirrordClusterPolicy`.
